Simplify the code that discovers the incoming message length and allocates a buffer for it.
We don't need seperate messages, just set the length before
recvmsg in both and track how big our allocated buffer is.

realloc() instead of malloc() to not lose memory.

'len' was just a temporary.

Hopefully this gets rid of a situation where iov.iov_len was garbage on the
stack and was being trusted.