rds-tools: update rds-info and man page
rds-info now outputs all of its information sources by default.  Each source is
prefaced by a description of its data and then headers for each column of its
data.

If options are specified then only the information associated with those
sources is displayed.  Multiple options may be specified.

To support all this the core of rds-info was refactored.  All the information
sources are enumerated in one struct array now.  The rest of the code works
from that array.

The message flag information is obsolete and is no longer output.

The man page is updated to reflect these changes.  It includes descriptions
of each of the columns for each data source.

Signed-off-by: Zach Brown <zach.brown@oracle.com>