rds-stress: Fix RTT estimation
This patch fixes a problem with the RTT estimate.

rds-stress uses a single sequence number space for all
packets, REQ and ACK. But the code recording the send time
stamp, and the code computing the RTT, assume that sequence
numbers correspond to REQ packets.

Assume we are sending the following 5 packets, then the
send_time is recorded in the following slots:

	REQ seqno=N		send_time[0]
	REQ seqno=N+1		send_time[1]
	REQ seqno=N+2		send_time[2]
	ACK seqno=N+3		-
	REQ seqno=N+4		send_time[0]

Now as ACKs come in, the sequence number on the incoming ACK (modulo the
queue depth) is used as an index into the send_time array.  This sequence
number has no relation with the sequence number originally used, and so
the RTT estimate is rather bogus.

The following patch introduces an "index" field in the packet header,
which identifies the send_time slot used.

Signed-off-by: Olaf Kirch <olaf.kirch@oracle.com>