rds-stress: suppress some duplicate warnings
When rds-stress detects that setsockopt(SNDBUF) will not give it as much
memory as it would like to have, every task will print the same warning.
Which is silly, especially when testing with 1024 tasks.

This patch suppresses these warnings for all tasks except the first
child process. It also makes these failures non-fatal.

Signed-off-by: Olaf Kirch <olaf.kirch@oracle.com>