rds-stress: fix check_parent()
The check_parent function is used to check whether the parent process
has exited. It does so by checking for the existence /proc/<ppid>.

However, this fails to work as designed when you use a perl script to
read the output of rds-stress, and the parent process dies. It will
stick around as a zombie, and /proc/<ppid> still exists. perl however
does not get a EOF from the pipe, and keeps on waiting for more input.

This patch makes check_parent() use getppid() instead, which changes the
moment the parent exits (as the process is reparented to the init process)

Signed-off-by: Olaf Kirch <olaf.kirch@oracle.com>