rds-stress: keep sending even when we see EAGAIN
There are two (major) reasons why sendmsg on an RDS socket can
fail:

 -	the send queue is full
 -	the destination port is congested

Unfortunately, in both cases RDS returns EAGAIN (well, it actually
returns EWOULDBLOCK in the latter, but that's the same error code
as EAGAIN).

So we cannot just break out of the send loop when we see EAGAIN,
but have to keep trying the other destinations as well.

Signed-off-by: Olaf Kirch <olaf.kirch@oracle.com>