rds-stress: move header building out of send_packet
In preparation of the rdma patches, move assembly of the
rds-stress header out of send_packet() and into the calling
functions.

It also changes the receive code to build the ACK header at the time
it receives the packet, rather than at the time the ACK is sent. This
is because with RDMA support, we need to remember a bunch of additional
values from the incoming packet anyway.

Signed-off-by: Olaf Kirch <olaf.kirch@oracle.com>