Stop leaking RDMA memory ranges
When we send a request containing a RDMA op, we allocate
a memory range (MR) and pass its key to the peer. If sending
the messages fails (usually because the send queue is full)
we leak the MR.

This patch adds yet another field to the task struct to
remember the MR key. I made this an array indexed by the
send_index, but it would also be sufficient to have just
one "uint64_t rdma_next_key" per task.

Signed-off-by: Olaf Kirch <olaf.kirch@oracle.com>