Use kernel ib_rds.h header
Instead of keeping a copy of all types and constants making up
the RDS API, we want to use the newly introduced ib_rds.h file from the
kernel source. For now, we have to avoid a build dependency on the
OFED kernel source though, so we still have to copy the API file.

Signed-off-by: Olaf Kirch <olaf.kirch@oracle.com>