Always fence RDMA reads
On the receiving end, the RDS message immediately following a RDMA READ
may arrive before the READ has completed. However, when we receice
the message, we will tear down the MR, so bad things will happen.

For this reason, we always fence RDMA READs in rds-stress. Other apps
may decide to take a different approach.

Signed-off-by: Olaf Kirch <olaf.kirch@oracle.com>