rds-stress: improvements for gathering perf data
When doing automated performance runs, it's nice to generate output
in a format that's easy to parse in a script. This script adds an
option "--show-perfdata" that does this.

At the same time, we gather additional statistics such as RDS
performance counters, and system stats such as interrupt count,
system times (system, user, iowait etc).

Signed-off-by: Olaf Kirch <olaf.kirch@oracle.com>