Fix RDS congestion monitoring
The RDS congestion monitoring code tries to help applications deal with
remote congestion more efficiently. If enabled, an application that tried
to send to a congested port will receive a notification as soon as the
port becomes uncongested again. For efficiency reasons, the application
isn't given a complete 8K congestion bitmap, but a 64bit mask that
represents the ports having changed, with port N being represented by
(1 << (port % 64))

The macro used to translate port numbers to the mask bit shifted integer
1, not 1ULL, resulting in undefined behavior when (port % 64) >= 32

Signed-off-by: Olaf Kirch <olaf.kirch@oracle.com>