rds-info: Extend IB connection info to include settings
RDS sysctl parameters such as the maximum number of WRs
on the send and receive queue are applied when a connection
is established, and some are further modified based on the
device capabilities. Hence, current sysctl values do not necessarily
reflect what is being used for a particulr RDS/IB connection.

The latest RDS kernel code extends struct rds_info_ib_connection to
include per-connection settings:

 -      max number of send and receive WRs
 -      max number of send S/G entries
 -      for RDMA, the maximum number of FMRs used
        in the RDMA pool, and the maximum size of a FMR.

This patch modifies rds-info to display these settings when
the new -v option is used.

Signed-off-by: Olaf Kirch <olaf.kirch@oracle.com>