accidental struct size change breaks connection flags
I am an idiot and thought it would be a good idea to change an array
size that was part of a struct exposed to userspace. The unfortunate
result of this is that old rds and rds-tools versions return a struct
1 byte shorter than what recent rds-tools expects.

Luckily rds-tools was written not to choke if the returned struct was
larger than expected, and the only affected data is the flags field after
the transport name. This hacks rds-info -n to also check the last byte
of the transport field for flags, in addition to conn.flags.

Signed-off-by: Andy Grover <andy.grover@oracle.com>