rds-stress: graceful stopping when using -T
Previously, using the -T option caused all child processes to be killed
when the time had elapsed. This patch gives them a second to stop sending,
so that when they are finally killed, there will be no outstanding traffic.
This should help the connection bounces caused when using RDMA and
terminating rds-stress.

Signed-off-by: Andy Grover <andy.grover@oracle.com>