rds-tools: rds-info updated to report information about IPv6 address
rds-info is updated to retrieve both IPv4 and IPv6 connection
information from kernel RDS.  By default, only IPv4 RDS connection
information is displayed.  If "-a" option is used, both IPv4 and IPv6
RDS connection information is displayed.  If the updated rds-info is
executed in a system running an RDS module which does not support IPv6,
only IPv4 information is displayed, with or without the "-a" option.

Orabug: 26646596

Signed-off-by: Ka-Cheong Poon <ka-cheong.poon@oracle.com>
Reviewed-by: Mukesh Kacker <mukesh.kacker@oracle.com>
Reviewed-by: Avinash Repaka <avinash.repaka@oracle.com>
Reviewed-by: Qing Huang <qing.huang@oracle.com>
Acked-by: Santosh Shilimkar <santosh.shilimkar@oracle.com>
Reviewed-by: Håkon Bugge <haakon.bugge@oracle.com>