rds-tools: rds-stress should work with IPv6 address
The command rds-stress sends a struct options to its peer.  And this
struct contains an IP address field which is a 32 bit integer.  To
work with IPv6 address, a new struct options_v6 is introduced to send
IPv6 address over.  To ensure backward compatibility, this new struct
is only sent to peer if the peer address is an IPv6 address.

Orabug: 26646596

Signed-off-by: Ka-Cheong Poon <ka-cheong.poon@oracle.com>
Reviewed-by: Mukesh Kacker <mukesh.kacker@oracle.com>
Reviewed-by: Avinash Repaka <avinash.repaka@oracle.com>
Reviewed-by: Qing Huang <qing.huang@oracle.com>
Acked-by: Santosh Shilimkar <santosh.shilimkar@oracle.com>