spec: Add "vos" to RPM release number
To disambiguate the RPM file names from builds of upstream pacakges
without Oracle-specific extensions, we add "vos" to the Release tag in
the RPM spec file.

At this time, this package is identical (source code and build
requirements) for all flavors of RDMA userpace packages in which it is
used (ofed-vos and rdma-vos) however, for consistency, we add the "vos"
token to its release to match the rest of the rdma-vos packages.

Orabug 28550856

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>