spec: Add epoch to package versioning
This change helps to ensure that upgrading from oracle-ofed-release to
oracle-rdma-release will work. See bug 29873464 for additional context.

The ofed-release to rdma-release update works for ofed-release versions
<= v1.0.0-48 but not for v1.0.0-49 or later. This is because v1.0.0-48
has rds-tools = 2.1.1-2.el7uek4, v1.0.0-49 and later have rds-tools =
2.2.0-1.el7uek4, and rdma-release has rds-tools = 2.2.0-1.el7ora.

The reason this does not work is because rds-tools in rdma-release does
not update rds-tools in ofed-release > v1.0.0-48

    rdma-release        ofed-release
    v2.2.0-1.el7ora  <  2.2.0-1.el7uek4

This would not have been a problem if the token in the Release tag of
the rdma-release packages had been left as "vos" and not changed to
"ora". Part of the reason for using "vos" was that it is greater than
"uek4" which is used for the ofed-release packages. This meant that even
when two packages have the same <version>-<release> the rdma-release
version would always be greater and update the ofed-release version. For
example:

    rdma-release        ofed-release
    v2.2.0-1.el7vos  >  2.2.0-1.el7uek4

because "vos" > "uek4".

This was particularly useful because it meant that we did not need to
introduce the use of Epoch into packages such as rds-tools for which (1)
the code is identical for ofed-release and rdma-release; (2) we
are the upstream (i.e., our version will always have a higher version
then the RHEL version); and (3) the rdma-release version is always
higher than the ofed-release version. Because of the forced change from
"vos" to "ora" (3) is no longer true.

To ensure that any version of ofed-release can be updated to
rdma-release we will add "Epoch: %{uek5epoch}" which is "Epoch: 5".

The rds-tools for ofed-release has implicit "Epoch: 0" but we can
optionally change it to "Epoch: %{uek4epoch}" which is "Epoch: 4" to be
consistent with all of the other ofed-release packages which use an
explicit Epoch.

Orabug: 29921620

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>