rds-tools: Allow rds-ping to set the number of RDS sockets used
Currently, rds-ping uses 8 RDS sockets to ping a remote host.  Though
this may be useful from an RDS testing perspective, it is not
necessary for the purpose of testing reachability.  A new option "-n"
is added allowing a user to specify the number of sockets used.  Valid
range is 1 to 32.  If "-n" is not specified but "-c x" is used to
specify the number of pings to sent where x < 8, the number of sockets
created is set to x as the other sockets created by default will not
be used.

Orabug: 30359242

Signed-off-by: Ka-Cheong Poon <ka-cheong.poon@oracle.com>
Acked-by: Santosh Shilimkar <santosh.shilimkar@oracle.com>
Reviewed-by: Håkon Bugge <haakon.bugge@oracle.com>

---
v1 -> v2
 - Changed to take "-c" option into account in socket creation as
   suggested by Håkon Bugge.

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>