rds-stress: SIGSEGV on OL8 when running IPv4
encode_hdr() and decode_hdr() need to use isv6 to determine the size
of the destination buffer. Otherwise, in the case of IPv4, the functions
will zero out 33 additional bytes.

This bug does not manifest in OFED, but it is still a bug all the same
and is being fixed.

Orabug: 30935289

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>
Reviewed-by: Håkon Bugge <haakon.bugge@oracle.com>