rds-tools: Update URL and Source tags in the spec
The "URL" and "Source0" tag values are out of date.

The upstream source for this package is now GitHub.

There is only one source file so use "Source" instead of "Source0".
The value of "Source" is set to the tarball name and not the full URL to
the tarball archive on GitHub. This is because we maintain a single spec
for two branches of the package (ofed-vos and rdma-vos) and do not yet
have a simple way (using Oracle's internal build system) to pass the
version into the spec file to generate the correct URL. For example:

github.com/oracle/rds-tools/archive/ofed-vos/rds-tools-2.3.1-1.tar.gz

or

github.com/oracle/rds-tools/archive/rdma-vos/rds-tools-2.3.1-1.tar.gz

Orabug: 30847018

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>