rds-info: Display destination qp number
This patch provides the user-space modifications required to display the
destination qp number corresponding to an rds connection. Essential changes
are made in the kernel to extract the destination qp number of a connection
and pass it to user-space. Invalid qp number is being provided the value of
-1 instead of 0 which was being used earlier, as 0 corresponds to a special
qp.

Orabug: 31940569

Suggested-by: Rama Nichanamatlu <rama.nichanamatlu@oracle.com>
Signed-off-by: Praveen Kumar Kannoju <praveen.kannoju@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>