rds-info: Add pid and comm fields to rds-info
Currently, given a rds_sock struct, there is no direct way to get to the
process that created the socket. We can use inode of rds_sock and output
of lsof to get to the process. This becomes tedious when the system
crashes and we do not have lsof output.

A kernel change was recently included to track and export the pid of
process that created rds_sock to user space. This is the user space part
of the change to display the pid and comm with 'rds-info -k' command.
This change does not break the compatibility if the kernel support for
pid is missing.

Orabug: 33504953

Signed-off-by: Rohit Nair <rohit.sajan.kumar@oracle.com>
Reviewed-by: Manjunath Patil <manjunath.b.patil@oracle.com>