rds-info: Fixed missing documentation for rds-info -I
A previous patch added the dst_qp_num as a displayable field to the
functionality of rds-info -I, but did not add the corresponding
documentation to the man page.

This patch modifies the code slightly to display a zero value for the
field instead of not printing the field when running on an older kernel
that does not support the field. It also documents these fields in the
man page.

Orabug: 33771194

Fixes: a437a5b1d750 ("rds-info: Display destination qp number")
Signed-off-by: Rohit Nair <rohit.sajan.kumar@oracle.com>
Reviewed-by: Manjunath Patil <manjunath.b.patil@oracle.com>