rds-tools: Sync with kernel rds.h version use of annotated types
Orabug: 34466041

The rds.h header file was originally copy of include/uapi/linux/rds.h from
the UEK kernel source. We should try to make it as close a copy of that header
file once again and then build using the kernel header file instead of
managing two copies of the same file. The userspace copy of the header file
should be modified to use the annotated types.

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>
Reviewed-by: Devesh Sharma <devesh.s.sharma@oracle.com>