rds-info: Fix strcasestr() implicit declaration warning
Orabug: 34484319

The fix in 7be4f24c ("rds-info: extend rds-info -Iv functionality")
introduced the use of the strcasestr() function. It is defined in
string.h, but must be preceded by "#define _GNU_SOURCE" as it is a
non-standard extension. Without that, a warning about an implicit
declaration of the function is printed to stderr.

warning: implicit declaration of function 'strcasestr'

Fixes: 7be4f24cffc1 ("rds-info: extend rds-info -Iv functionality")
Signed-off-by: Mark Haywood <mark.haywood@oracle.com>
Reviewed-by: Gerd Rausch <gerd.rausch@oracle.com>
Reviewed-by: Sharath Srinivasan <sharath.srinivasan@oracle.com>