rds-info: Move the conn_drop_reasons array from rds.h to rds-info.c
Orabug: 34466041

The conn_drop_reasons array was added so that rds-info could report
text explanations for connection drops. Unfortunately, it was added
to rds.h. We would like to remove rds.h from the rds-tools project
as it was orignally a copy of include/uapi/linux/rds.h from the UEK
kernel source. We should try to make it as close a copy of that
header file once again and then build using the kernel header file
instead of managing two copies of the same file.

Fixes: eb793ecd0fef ("rds-tools: Added '-p' option to rds-info to show paths")
Signed-off-by: Mark Haywood <mark.haywood@oracle.com>
Reviewed-by: Devesh Sharma <devesh.s.sharma@oracle.com>