rds-stress: Add new json options exchange logic for rds-stress
Orabug: 34979190

The current options negotiation in rds-stress is quite rigid and does
not allow the addition of new options due to it's implementation.

With this situation, the motivation was to create a new options
negotiation mechanism that can be robust and flexible. The new mechanism
should allow the addition of new options and the removal of options in
the future. To achieve this, we exchange the options in a json encoded
string format that can be used to fill in the options struct and not be
limited as we have been in the past.

This patch adds a new option exchange mechanism that utilizes json-c.
This patch supports backwards compatibility by using the legacy option
exchange method with previous versions of rds-stress.

Signed-off-by: Rohit Nair <rohit.sajan.kumar@oracle.com>
Reviewed-by: Manjunath Patil <manjunath.b.patil@oracle.com>