rds-stress: Use port[0] for rds-stress reset
Currently rds-stress uses the same default port[4000] when resetting a
connection, though binding to a particular port is not necessary. This
can cause a bind error() for the rds connection reset attempt if the
port number is already in use.

To avoid this possible error, we use port[0] when performing a reset.

Orabug: 35156783

Signed-off-by: Rohit Nair <rohit.sajan.kumar@oracle.com>
Reviewed-by: Gerd Rausch <gerd.rausch@oracle.com>
Reviewed-by: Arumugam Kolappan <aru.kolappan@oracle.com>