rds-stress: Fix initialization of rtt_threshold
Orabug: 35284851

rtt_threshold is a uint64_t data type. When rds-stress is invoked,
rtt_threshold should be initialized to the largest value for that
data type. And if rds-stress is invoked with the --show-outliers
option, then rtt_threshold should be set a value no larger than
that data type can handle.

Suggested-by: Håkon Bugge <haakon.bugge@oracle.com>
Signed-off-by: Mark Haywood <mark.haywood@oracle.com>
Reviewed-by: William Kucharski <william.kucharski@oracle.com>