rds-info: Add option to report information for all network namespaces
Orabug: 34270854

The addition of RDS/TCP support for non-initial network namespaces
creates the need for obtaining RDS information from such namespaces.

While it is possible to get the desired information by executing the
namespace-unaware rds-info(1) directly in other network namespaces,
this approach requires extra effort from the user to determine the
identifier for each additional namespace and to run rds-info(1)
separately in each of them.

In a container environment, one can conveniently run the
namespace-unaware rds-info(1) in a container (with its own network
namespace) without the need to determine the network namespace
identifier.  However, rds-info(1) must be installed in each container
of interest, which conflicts with the common desire to minimize
container environments, and rds-info(1) must again be run separately
in each container.

Thus, this patch introduces an rds-info(1) feature for iterating over
all network namespaces and printing RDS information for each of them.
Note that rds-info(1) requires additional privileges to access other
namespaces.

Signed-off-by: Greg Jumper <greg.jumper@oracle.com>
Reviewed-by: Mark Haywood <mark.haywood@oracle.com>