Revert "rds-ping: add transport option"
Orabug: 35658752

This reverts commit aa9fe068e0499c8758a02a5876ef050511a709b7.

This patch was reverted because the addition of the transport
option is unnecessary and encourages the use of an unsupported
configuration where both rds_rdma and rds_tcp modules are loaded.

Conflicts:
        rds-ping.1
        rds-ping.c

Conflicts had to be resolved as a result of the subsequent addition
of whitespace changes introduced by f01bc028 ("rds-ping: Add
missing '-i'option to rds-ping usage") and d5216aec ("man: Add
missing '-Q' option to rds-ping(1)").

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>
Reviewed-By: Devesh Sharma <devesh.s.sharma@oracle.com>