rds-stress: Add support for option "--always-bursty"
Orabug: 35792064

We add an option "--always-bursty" to "rds-stress" that retains
the bursty behavior, even in cases where the sendmsg-path is slower
than the recvmsg-path.

1) For as long as messages are to be sent (pending<req_depth),
   it'll stay in BURST_MODE_SEND
2) For as long as messages are to be received (pending>0),
   it'll stay in BURST_MODE_RECV

Also, exit with an error if the socket buffer (SO_SNDBUF, SO_RCVBUF)
is insufficient to hold one batch of messages, as we no longer mix modes.

Background:

Batch processing in rds-stress causes paradoxical behavior.

The following simplified pseudo-code illustrates the current
working mechanism of "rds-stress":
--------%<--------%<--------%<--------%<--------%<--------%<--------
while (testing) {
    poll(...);
    while (recvmsg(...)); /* terminates by -EAGAIN) */
    while (data_to_send && POLL_OUT) sendmsg(...);
}
--------%<--------%<--------%<--------%<--------%<--------%<--------

This makes "rds-stress" alternate between receive-all and send-all
batch processing.

That approach adds collateral delay to the send- and receive-path,
because a message that's ready to be sent has to wait until the
receive-phase has finished, and likewise a message that's ready to be
received won't be received until the send-phase has finished.

With roughly equal "sendmsg" and "recvmsg" performance, the first message
received in a batch with N messages, has to wait for "(N - 1) * recvmsg_cost"
until the code transitions to the send-phase.

Only if the "sendmsg" and "recvmsg" costs are unequal, will the batch/bursty
traffic pattern transition to an interleaved one, with less penalty
while waiting for either of the send- or receive-mode to finish.

For example, if the "sendmsg" cost is increased enough (slowed down)
to subsequently increase the distance (time-wise) between two consecutive
messages sent, the receive-batch can reach -EAGAIN sooner and before
the entire batch was received, and thus start pushing messages out again.
This will utilize the link/pipeline more efficiently.

Unfortunately, rds-stress transitions from batch-processing to interleaved
in an uncontrolled manner, and unknown to an observer.
This makes the existing rds-stress tool unsuitable to be used for reliable
performance measurements.

E.g. a slowdown in the send-path carries the risk of being misinterpreted
as a net performance-gain, since it likely makes rds-stress report better numbers.

Experiments have shown that an injection of a 75usec delay into the send-path
("rds_copy_to_user_delay") improved performance by ~75%.
An injection of a 20usec delay into the receive-path ("rds_copy_to_user_delay")
improved performance by ~92% (almost double).

For obvious reasons, making the code slower isn't the right path forward.

But unfortunately, the argument that rds-stress showed better numbers was used
in the past to justify various code-additions to the send-path.

Signed-off-by: Gerd Rausch <gerd.rausch@oracle.com>
Reviewed-by: Mark Haywood <mark.haywood@oracle.com>