rds-stress: Exchange the rds-inq value using json
Orabug: 35800936

When the rds-inq support was added, it should have been the first option
to use the newly supported json option exchange. This means it should
have frozen the size of the old option structure exchange. It also means
that it should have enabled the json exchange by enabling use_json.

Fixes: fd833d3067a4 ("rds-stress: Add new json options exchange logic for rds-stress")
Fixes: 20f736d4be15 ("Add a flag to rds-stress and enable rds-inq capability")
Signed-off-by: Mark Haywood <mark.haywood@oracle.com>
Reviewed-by: Rohit Nair <rohit.sajan.kumar@oracle.com>