rds-stress: Synchronize burst-mode with other processes
Orabug: 35861956

Add option "--sync-always-bursty" to synchronize the
send-only and receive-only modes with other processes
via a System V semaphore.

Without this option, the send-only and receive-only
modes of processes are not synchronized, and can
drift from each other, especially with multiple invocations
of "rds-stress".

All processes sending at the same time, and subsequently
hitting upon bottlenecks at the same time, is a different
performance-test than processes sending at slightly
different times.

If that distinction is uncontrolled and left to chance,
performance measurements are left to chance as well.

We create two System V semaphores, one for each burst-mode.
Each process in a mode increments the mode-counter by one.
When it's time to switch the mode, we wait for all processes
having exited the old mode (i.e. semaphore-value == 0),
before we enter the new mode.

Signed-off-by: Gerd Rausch <gerd.rausch@oracle.com>
Reviewed-by: Mark Haywood <mark.haywood@oracle.com>